Refactor scope (#3652)
* refactor Session.job_name2current_scope_ to Session.scope_stack_

* Seperate scope related functions from Session to scope_util.py

* revert change
