+++ oneflow/python/eager/interpreter_callback.py
@@ -22,14 +22,14 @@ import oneflow.core.job.job_conf_pb2 as job_conf_pb
 import oneflow.core.job.placement_pb2 as placement_pb
 from google.protobuf import text_format
 import oneflow.python.eager.blob_register as blob_register_util
-import oneflow.python.framework.compiler as compiler
+import oneflow.python.framework.scope_util as scope_util
 import oneflow.python.eager.vm_util as vm_util
 
 
 def MakeScopeSymbol(job_conf_str, parallel_conf_str, is_mirrored):
     job_conf = text_format.Parse(job_conf_str, job_conf_pb.JobConfigProto())
     parallel_conf = text_format.Parse(parallel_conf_str, placement_pb.ParallelConf())
-    return compiler.MakeInitialScope(
+    return scope_util.MakeInitialScope(
         job_conf, parallel_conf.device_tag, list(parallel_conf.device_name), is_mirrored
     ).symbol_id
